Updated the schedule builder interface.
Included two buttons where one can save the schedule and the other can display the schedule made previously.

Added a schedule.json file that will store the scheduling hours information. This file will be used to access the schedule in the employee perspective.

Added two menu items that can clear the schedule view or delete the entire schedule and start from the beginning.